Protect against overflow when computing metrics.
Reported by Gabriel Kerneis.

+++ message.c
@@ -974,7 +974,7 @@ flushupdates(struct network *net)
                     continue;
                 if(!route_interferes(route, net)) {
                     metric =
-                        route->refmetric +
+                        (int)route->refmetric +
                         (diversity_factor * route->cost / 128) / 256 +
                         route->add_metric;
                     metric = MAX(metric, route->refmetric + 1);
